UFS data wiping failure avoidance
I have noticed that on ufs if the file size is too small, less than optimal block size, if you force it to be equals to block size, the filesystem will discard the current inode and write the new bloated data block to a new inode. The problem with this strategy is that the old inode will not be really erased from disk. Forcing blkpad be equals to zero will avoid this harmful strategy when considering privacy.